% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.tmb_ml_analysis}
\alias{as.tmb_ml_analysis}
\title{Coerce to an TMB ML Analysis Object}
\usage{
as.tmb_ml_analysis(x, ...)
}
\arguments{
\item{x}{object to coerce.}

\item{...}{Unused.}
}
\description{
Coerce to an TMB ML Analysis Object
}
