% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.tmb_analysis}
\alias{augment.tmb_analysis}
\title{Augment TMB Analysis}
\usage{
\method{augment}{tmb_analysis}(x, conf_int = FALSE, conf_level = 0.95, ...)
}
\arguments{
\item{x}{The tmb_analysis object.}

\item{conf_int}{A flag indicating whether to calculate confidence intervals (using likelihood profiling).}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{...}{Unused.}
}
\value{
The data with the fitted and residual values.
}
\description{
Augments the data of a TMB analysis object.
}
\seealso{
\code{\link[broom]{augment}}.
}

