% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.tmb_analysis}
\alias{coef.tmb_analysis}
\title{Coef TMB Analysis}
\usage{
\method{coef}{tmb_analysis}(object, terms = "fixed", conf.int = FALSE,
  conf.level = 0.95, ...)
}
\arguments{
\item{object}{The tmb_analysis object.}

\item{terms}{A string of the terms to tidy. Permitted values are 'all', 'fixed',
'random' and 'report'.}

\item{conf.int}{A flag indictating whether to calculate confidence intervals.}

\item{conf.level}{A number specifying the confidence level. By default 0.95.}

\item{...}{Unused.}
}
\description{
Coefficients for a TMB analysis.
}

