% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converged.R
\name{converged.tmb_analysis}
\alias{converged.tmb_analysis}
\title{Is Converged}
\usage{
\method{converged}{tmb_analysis}(x, mcmc = FALSE,
  rhat = getOption("mb.rhat", 1.1), ...)
}
\arguments{
\item{x}{The object to test convergence for.}

\item{mcmc}{A flag specifying whether to get the mcmc draws as opposed to ml estimates.}

\item{rhat}{A number specifying the rhat threshold.}

\item{...}{Unused.}
}
\description{
Is Converged
}

