% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converged.R
\name{converged.tmb_ml_analysis}
\alias{converged.tmb_ml_analysis}
\title{Is Converged}
\usage{
\method{converged}{tmb_ml_analysis}(x, ...)
}
\arguments{
\item{x}{The object to test convergence for.}

\item{...}{Unused.}
}
\description{
Is Converged
}
