% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.tmb_analysis}
\alias{estimates.tmb_analysis}
\title{Estimates}
\usage{
\method{estimates}{tmb_analysis}(object, fixed = TRUE, mcmc = FALSE, ...)
}
\arguments{
\item{object}{The mb_analysis object.}

\item{fixed}{A flag specifying whether fixed or random terms.}

\item{mcmc}{A flag specifying whether to get the mcmc draws as oppsed to ml estimates.}

\item{...}{Not used.}
}
\description{
estimates
}

