% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance.tmb_analysis}
\alias{glance.tmb_analysis}
\title{Glance TMB Analysis}
\usage{
\method{glance}{tmb_analysis}(x, ...)
}
\arguments{
\item{x}{The tmb_analysis object to convert to a single-row data frame.}

\item{...}{Unused.}
}
\description{
A single row data frame of summary model statistics.
}
\seealso{
\code{\link[broom]{glance}}.
}

