% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{inits}
\alias{inits}
\title{Inits}
\usage{
inits(x)
}
\arguments{
\item{x}{The object.}
}
\value{
The initial values as a named list.
}
\description{
Gets a named list of the initial values for an object.
}

