% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.tmb_code}
\alias{is.tmb_code}
\title{Is a TMB Code}
\usage{
is.tmb_code(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'tmb_code'
}
