% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.tmb_mcmc_analysis}
\alias{is.tmb_mcmc_analysis}
\title{Is a TMB MCMC Analysis}
\usage{
is.tmb_mcmc_analysis(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'tmb_mcmc_analysis'
}

