% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.tmb_ml_analysis}
\alias{is.tmb_ml_analysis}
\title{Is a TMB ML Analysis}
\usage{
is.tmb_ml_analysis(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'tmb_ml_analysis'
}
