% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.tmb_model}
\alias{is.tmb_model}
\title{Is a TMB Model?}
\usage{
is.tmb_model(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests wether x is an object of class 'tmb_model'
}

