% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.tmb_ml_analysis}
\alias{logLik.tmb_ml_analysis}
\title{Log-Likelihood}
\usage{
\method{logLik}{tmb_ml_analysis}(object, ...)
}
\arguments{
\item{object}{The tmb_analysis object.}

\item{...}{unused.}
}
\description{
Log-Likelihood for a TMB ML analysis.
}
