% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-code.R
\name{model_code}
\alias{model_code}
\title{Model Code}
\usage{
model_code(x)
}
\arguments{
\item{x}{The object.}
}
\value{
The model code as a character vector for an object
}
\description{
Gets the model_code for an object.
}

