% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-new.R
\name{modify_new}
\alias{modify_new}
\title{Modify New}
\usage{
modify_new(x)
}
\arguments{
\item{x}{The object.}
}
\value{
The modify_new function.
}
\description{
Gets the modify_new function for an object.
}

