% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-new.R
\name{modify_new_data}
\alias{modify_new_data}
\title{Modify New Data}
\usage{
modify_new_data(x)
}
\arguments{
\item{x}{The object.}
}
\value{
The modify_new_data function.
}
\description{
Gets the modify_new_data function for an object.
}

