% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-code.R
\name{new_code}
\alias{new_code}
\title{New Code}
\usage{
new_code(x)
}
\arguments{
\item{x}{The object.}
}
\value{
The predict code as a character vector for an object
}
\description{
Gets the new_code for an object.
}

