% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Parameters}
\usage{
parameters(x, terms = "fixed")
}
\arguments{
\item{x}{The object.}

\item{terms}{A string of the terms to get the parameter names for.}
}
\value{
The parameter names as a character vector.
}
\description{
Gets the fixed or random or both (and in the case of analysis report and all)
parameter names for an object.
}

