% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.tmb_analysis}
\alias{predict.tmb_analysis}
\title{Predict}
\usage{
\method{predict}{tmb_analysis}(object, new_data = data_set(object),
  term = "prediction", new_expr = NULL, select_new_data = NULL,
  modify_new_data = NULL, ...)
}
\arguments{
\item{object}{The tmb_analysis object.}

\item{new_data}{The data frame to calculate the predictions for.}

\item{term}{A string of the term in new_expr.}

\item{new_expr}{A string of the R code specifying the predictive relationships.}

\item{select_new_data}{A named list specifying the columns in a new data frame to select and their associated classes and values.}

\item{modify_new_data}{A single argument function to modify the new_data (in list form) immediately prior to the predictions.}

\item{...}{Unused.}
}
\value{
The new data with the predictions.
}
\description{
Calculate predictions.
}

