% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-code.R
\name{predict_code}
\alias{predict_code}
\title{Predict Code}
\usage{
predict_code(x)
}
\arguments{
\item{x}{The object.}
}
\value{
The predict code as a character vector for an object
}
\description{
Gets the predict_code for an object.
}

