% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile}
\alias{profile}
\title{Profile}
\usage{
profile(object, new_data, profile_expr = NULL, back_transform = identity,
  conf_level = 0.95)
}
\arguments{
\item{object}{The tmb_analysis object to profile.}

\item{new_data}{A data.frame to use for profile.}

\item{profile_expr}{The profiling expression. Must be linear.}

\item{back_transform}{A single argument function specifying the transformation for the estimate and confidence intervals.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}
}
\value{
The new data with estimates and lower and upper confidence intervals.
}
\description{
Profile
}

