% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-effects.R
\name{random_effects}
\alias{random_effects}
\title{Random Effects}
\usage{
random_effects(x)
}
\arguments{
\item{x}{The object.}
}
\value{
The parameters as a list.
}
\description{
Gets the random effects parameters for an object.
}

