% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.tmb_analysis}
\alias{residuals.tmb_analysis}
\title{Residuals}
\usage{
\method{residuals}{tmb_analysis}(object, ...)
}
\arguments{
\item{object}{The tmb_analysis object.}

\item{...}{Unused.}
}
\value{
The data with the residual values.
}
\description{
Extract residual values for a TMB analysis.
}

