% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-new-data.R
\name{select_new_data}
\alias{select_new_data}
\title{Select New Data}
\usage{
select_new_data(x)
}
\arguments{
\item{x}{The object.}
}
\value{
The select_new_data object
}
\description{
Gets the select_new_data object for an object.
}

