% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.tmb_analysis}
\alias{tidy.tmb_analysis}
\title{Tidy TMB Analysis}
\usage{
\method{tidy}{tmb_analysis}(x, terms = "fixed", conf.int = FALSE,
  conf.level = 0.95, ...)
}
\arguments{
\item{x}{The tmb_analysis object to tidy.}

\item{terms}{A string of the terms to tidy. Permitted values are 'all', 'fixed',
'random' and 'report'.}

\item{conf.int}{A flag indictating whether to calculate confidence intervals.}

\item{conf.level}{A number specifying the confidence level. By default 0.95.}

\item{...}{Unused.}
}
\description{
Tidys the coefficients of a TMB analysis object into a data frame.
}
\seealso{
\code{\link[broom]{tidy}}.
}

