% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-analysis.R
\name{tmb_analysis}
\alias{tmb_analysis}
\title{TMB Analysis}
\usage{
tmb_analysis(data_set, model, beep = TRUE, debug = FALSE, ...)
}
\arguments{
\item{data_set}{The data frame to analyse.}

\item{model}{The model to use for the analysis.}

\item{beep}{A flag indicating whether to beep on completion of the analysis.}

\item{debug}{A flag indicating whether to run in debug mode.}

\item{...}{Not used.}
}
\description{
Analyse a data frame using a TMB model.
}

