% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.tmb_analysis}
\alias{augment.tmb_analysis}
\title{Augment TMB Analysis}
\usage{
\method{augment}{tmb_analysis}(x, terms = c("fit", "residual"),
  conf_level = 0.95, ...)
}
\arguments{
\item{x}{The tmb_analysis object to augment.}

\item{terms}{The terms to augment the data_set with.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{...}{Unused.}
}
\description{
Augments the data_set of a TMB analysis object.
}
\seealso{
\code{\link[broom]{augment}}.
}

