% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.tmb_analysis}
\alias{coef.tmb_analysis}
\title{Coef TMB Analysis}
\usage{
\method{coef}{tmb_analysis}(object, fixed = TRUE, include_constant = TRUE,
  mcmc = FALSE, conf_level = 0.95, ...)
}
\arguments{
\item{object}{The mb_analysis object.}

\item{fixed}{A flag specifying whether fixed or random terms.}

\item{include_constant}{A flag specifying whether to include constant terms.}

\item{mcmc}{A flag specifying whether to get the mcmc draws as oppsed to ml estimates.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{...}{Not used.}
}
\value{
A tidy tibble of the coefficient terms.
}
\description{
Coefficients for a TMB analysis.
}
\details{
The (95\%) \code{lower} and \code{upper} confidence intervals are
the \code{estimate} +/- 1.96 * \code{std.error}.
}

