% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.tmb_analysis}
\alias{coef.tmb_analysis}
\title{Coef TMB Analysis}
\usage{
\method{coef}{tmb_analysis}(object, terms = "fixed", conf_int = TRUE,
  conf_level = 0.95, ...)
}
\arguments{
\item{object}{The tmb_analysis object.}

\item{terms}{A string of the terms to tidy. Permitted values are 'all', 'fixed',
'random' and 'report'.}

\item{conf_int}{A flag specifying whether to calculate (approximate) confidence intervals.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{...}{unused.}
}
\description{
Coefficients for a TMB analysis.
}
\details{
The \code{statistic} is the z value.
The \code{p.value} is \code{Pr(>|z^2|)}.
The (95\%) \code{lower} and \code{upper} confidence intervals are
the \code{estimate} +/- 1.96 * \code{std.error}.
}

