% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.tmb_analysis}
\alias{fitted.tmb_analysis}
\title{Fitted Values}
\usage{
\method{fitted}{tmb_analysis}(object, term = "fit", conf_level = 0.95, ...)
}
\arguments{
\item{object}{The tmb_analysis object.}

\item{term}{A string of the terms to tidy.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{...}{unused.}
}
\value{
The data_set with the fitted values.
}
\description{
Extract fitted values for a TMB analysis.
}

