% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.tmb_analysis}
\alias{fitted.tmb_analysis}
\title{Fitted Values}
\usage{
\method{fitted}{tmb_analysis}(object, ...)
}
\arguments{
\item{object}{The tmb_analysis object.}

\item{...}{Unused.}
}
\value{
The data with the fitted values.
}
\description{
Extract fitted values for a TMB analysis.
}

