% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.tmb_analysis}
\alias{predict.tmb_analysis}
\title{Predict}
\usage{
\method{predict}{tmb_analysis}(object, new_data = data_set(object),
  term = "prediction", new_code = new_code(object),
  modify_new_data = modify_new_data(object), profile = FALSE,
  conf_level = 0.95, ...)
}
\arguments{
\item{object}{The tmb_analysis object.}

\item{new_data}{The data frame to calculate the predictions for.}

\item{term}{A string of the term in new_code.}

\item{new_code}{A string of the R code specifying the predictive relationships.}

\item{modify_new_data}{A single argument function to modify the new_data (in list form) immediately prior to the predictions.}

\item{profile}{A flag indicating whether to calculate the predictions using likelihood profiling.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{...}{unused.}
}
\value{
The newdata with the predictions.
}
\description{
Calculate predictions.
}

