% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.tmb_analysis}
\alias{predict.tmb_analysis}
\title{Predict}
\usage{
\method{predict}{tmb_analysis}(object, new_data = data_set(object),
  term = "prediction", predict_expr = NULL, ...)
}
\arguments{
\item{object}{The tmb_analysis object.}

\item{new_data}{The data frame to calculate the predictions for.}

\item{term}{A string of the term in predict_expr.}

\item{predict_expr}{A string of R code specifying the predictive relationships.}

\item{...}{Unused.}
}
\value{
The new data with the predictions.
}
\description{
Calculate predictions.
}

