% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.tmb_analysis}
\alias{residuals.tmb_analysis}
\title{Residuals}
\usage{
\method{residuals}{tmb_analysis}(object, conf_int = FALSE,
  conf_level = 0.95, ...)
}
\arguments{
\item{object}{The tmb_analysis object.}

\item{conf_int}{A flag indicating whether to calculate confidence intervals (using likelihood profiling).}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{...}{Unused.}
}
\value{
The data with the residual values.
}
\description{
Extract residual values for a TMB analysis.
}

