% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-analysis.R
\name{tmb_analysis}
\alias{tmb_analysis}
\title{TMB Analysis}
\usage{
tmb_analysis(data, model)
}
\arguments{
\item{data}{The data.frame to analyse.}

\item{model}{The tmb_model to analyse.}
}
\value{
An object of class tmb_analysis.
}
\description{
TMB Analysis
}

