% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{tmb_model}
\alias{tmb_model}
\title{TMB Model}
\usage{
tmb_model(model_code, inits, select = character(0), center = character(0),
  scale = character(0), random = character(0))
}
\arguments{
\item{model_code}{A string of the model template code.}

\item{inits}{A named list of initial values for all fixed and random parameters.}

\item{select}{A character vector or a named list specifying the columns to select (and in the case of a named list the associated classes and values).}

\item{center}{A character vector of the columns to center.}

\item{scale}{A character vector of the columns to scale (after centering).}

\item{random}{A character vector naming the random effects parameters.}
}
\value{
An object of class tmb_model.
}
\description{
TMB Model
}
\seealso{
\code{\link[datacheckr]{check_data}} \code{\link[rescale]{rescale}}
}

