% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{tmb_model}
\alias{tmb_model}
\title{TMB Model}
\usage{
tmb_model(model_code, parameters, select_data = NULL, random = NULL)
}
\arguments{
\item{model_code}{A string of the model template.}

\item{parameters}{A list of all parameter objects required by the user template.}

\item{select_data}{NULL or a character vector or a named list specifying the variables to pass to the analysis (and in the case of a named list the associated classes and values).}

\item{random}{A character vector specifying the random effects parameters.}
}
\value{
An object of class tmb_model.
}
\description{
If select is a named list then check_data2 is used to check the consistency of
the data.
}

