% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{tmb_model}
\alias{tmb_model}
\title{TMB Model}
\usage{
tmb_model(model_code, gen_inits, random_effects = list(),
  select_data = list(), center = character(0), scale = character(0),
  modify_data = identity, new_expr = character(0),
  profile_expr = character(0))
}
\arguments{
\item{model_code}{A string of the model template code.}

\item{gen_inits}{A single argument function taking the modified data and
returning a named list of initial values for all fixed parameters.
Missing random parameters are assigned the value 0.}

\item{random_effects}{A named list specifying the random effects and the associated factors.}

\item{select_data}{A named list specifying the columns to select and their associated classes and values.}

\item{center}{A character vector of the columns to center.}

\item{scale}{A character vector of the columns to scale (after centering).}

\item{modify_data}{A single argument function to modify the data (in list form) immediately prior to the analysis.}

\item{new_expr}{A string of R code specifying the predictive relationships.}

\item{profile_expr}{A string of R code specifying the linear relationships.}
}
\value{
An object of class tmb_model.
}
\description{
Creates TMB model.
}
\seealso{
\code{\link[datacheckr]{check_data}} \code{\link[rescale]{rescale}}
}

