% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{tmb_model}
\alias{tmb_model}
\title{TMB Model}
\usage{
tmb_model(model_code, inits, random_effects = character(0),
  select_data = character(0), center = character(0), scale = character(0),
  modify_data = function(x) x, new_code = character(0),
  select_new_data = select_data, modify_new_data = modify_data)
}
\arguments{
\item{model_code}{A string of the model template code.}

\item{inits}{A named list of initial values for all fixed and random parameters.}

\item{random_effects}{A character vector or a named list specifying of the random effects (and in the case of a named list the associated factors).}

\item{select_data}{A character vector or a named list specifying the columns to select (and in the case of a named list the associated classes and values).}

\item{center}{A character vector of the columns to center.}

\item{scale}{A character vector of the columns to scale (after centering).}

\item{modify_data}{A single argument function to modify the data (in list form) immediately prior to the analysis.}

\item{new_code}{A string of the R code specifying the predictive relationships.}

\item{select_new_data}{A character vector or a named list specifying the columns in a new data frame to select (and in the case of a named list the associated classes and values).}

\item{modify_new_data}{A single argument function to modify the new_data (in list form) immediately prior to the predictions.}
}
\value{
An object of class tmb_model.
}
\description{
Creates TMB model.
}
\seealso{
\code{\link[datacheckr]{check_data}} \code{\link[rescale]{rescale}}
}

