% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{tmb_model}
\alias{tmb_model}
\title{TMB Model}
\usage{
tmb_model(model_code, inits, random = character(0), select = character(0),
  center = character(0), scale = character(0), modify = function(x) x)
}
\arguments{
\item{model_code}{A string of the model template code.}

\item{inits}{A named list of initial values for all fixed and random parameters.}

\item{random}{A character vector or a named list specifying of the random effects (and in the case of a named list the associated factors).}

\item{select}{A character vector or a named list specifying the columns to select (and in the case of a named list the associated classes and values).}

\item{center}{A character vector of the columns to center.}

\item{scale}{A character vector of the columns to scale (after centering).}

\item{modify}{A single argument function to modify the data (in list form) immediately prior to the analysis.}
}
\value{
An object of class tmb_model.
}
\description{
Creates TMB model.
}
\details{
With the exception of \code{model_code} and \code{select} all arguments are sorted.
}
\seealso{
\code{\link[datacheckr]{check_data}} \code{\link[rescale]{rescale}}
}

