% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{tmb_model}
\alias{tmb_model}
\title{TMB Model}
\usage{
tmb_model(model_code, parameters, select_data = character(0),
  center = character(0), scale = character(0),
  random_effects = character(0))
}
\arguments{
\item{model_code}{A string of the model template.}

\item{parameters}{A list of all parameter objects required by the user template (both random and fixed).}

\item{select_data}{A character vector or a named list specifying the columns to select (and in the case of a named list the associated classes and values).}

\item{center}{A character vector of the columns to center.}

\item{scale}{A character vector of the columns to scale (after centering).}

\item{random_effects}{A character vector specifying the random effects parameters.}
}
\value{
An object of class tmb_model.
}
\description{
TMB Model
}
\seealso{
\code{\link[datacheckr]{check_data}} \code{\link[rescale]{rescale}}
}

