% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{tmb_model}
\alias{tmb_model}
\title{TMB Model}
\usage{
tmb_model(model_code, parameters)
}
\arguments{
\item{model_code}{A string of the model template.}

\item{parameters}{A list of all parameter objects required by the user template.}
}
\value{
An object of class tmb_model.
}
\description{
TMB Model
}

